some changes to the UI and functions
User login is added. When you load the page it asks you to enter username, if you don't enter username it's not gonna let you continue. once username is entered, you can add group, if group is not added it's not gonna let you add Item. When an item is created, it comes with a title, a checkbox to make visible (still need to work on that), a button that the user can select to lock the item to work on it (still needs work), the user can add text to the item